+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 edbee.lib:
 
+- ref #26, Changing a theme in the ThemeManager, now updates the ThemePointer used by the renderer. (Fixes corruption on theme reloading)
+- fix #25, Removed the 'memory ok :-D' message
 - fix #24, Clearing the undostack unregistered all controllers incorrectly
 - fix #23, Scope invaldating optimization wasn't working correctly. Removed optimization for now
 - fix #19, Workaround-hack for non-bmp-unicode character movement. (experimental)